requirements.txt
amqp==5.1.1
appdirs==1.4.4
asgiref==3.5.2
astroid==2.8.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.7.1
Automat==20.2.0
billiard==3.6.4.0
blis==0.7.9
cached-property==1.5.2
cachetools==4.2.4
catalogue==1.0.2
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
channels==4.0.0
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coloredlogs==14.3
colorthief==0.2.1
colour==0.1.5
constantly==15.1.0
cryptography==37.0.2
cymem==2.0.7
daphne==4.0.0
decorator==4.4.2
Deprecated==1.2.13
Django==4.0.4
django-appconf==1.0.5
django-cors-headers==3.11.0
django-multiselectfield==0.1.12
django-select2==7.10.0
django-taggit==3.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
docopt==0.6.2
filter==0.0.0.20200724
fs==2.4.16
google-api-core==1.32.0
google-api-python-client==2.58.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.0
gunicorn==20.1.0
h11==0.13.0
httplib2==0.20.4
humanfriendly==10.0
hyperlink==21.0.0
idna==2.7
imageio==2.21.0
imageio-ffmpeg==0.4.7
importlib-metadata==4.11.3
incremental==22.10.0
instaloader==4.5.5
instalooter==2.4.4
isort==5.9.3
kombu==5.2.4
lazy-object-proxy==1.6.0
Markdown==3.3.7
mccabe==0.6.1
more-itertools==8.14.0
moviepy==1.0.3
murmurhash==1.0.9
notifications==0.3.2
numpy==1.23.1
oauthlib==3.2.1
opencv-python==4.6.0.66
ordered-set==3.1.1
ordered-set-stubs==0.1.3
outcome==1.2.0
packaging==21.3
pandas==1.5.1
Pillow==9.2.0
pipreqs==0.4.11
plac==1.1.3
platformdirs==2.4.0
poetry-version==0.1.5
preshed==3.0.8
profanity==1.1
profanity-filter==1.3.3
proglog==0.1.10
prompt-toolkit==3.0.30
protobuf==3.14.0
psutil==5.9.2
psycopg2==2.9.4
psycopg2-binary==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
PyJWT==2.4.0
pylint==2.11.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.21.0
python-unsplash==1.1.0
pytube==12.1.0
pytz==2022.1
pyunsplash==1.0.0rc2
redis==3.5.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.15.100
schedule==0.6.0
selenium==4.3.0
service-identity==21.1.0
six==1.10.0
sniffio==1.2.0
sortedcontainers==2.4.0
spacy==2.3.8
sqlparse==0.4.3
srsly==1.0.6
tenacity==6.3.1
terminedia==0.4.2
thinc==7.4.6
toml==0.10.2
tomlkit==0.5.11
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
tweepy==4.10.1
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
uritemplate==3.0.1
urllib3==1.24.3
verboselogs==1.7
vine==5.0.0
wasabi==0.10.1
wcwidth==0.2.5
wrapt==1.13.2
wsproto==1.1.0
yarg==0.1.9
zipp==3.8.0
zope.interface==5.5.0
